Added H.264 support thanks to Matthias Schneider <ma30002000 yahoo de>. Thanks a lot !
Baseline Profile:
no B-frames
We make use of the baseline profile (which is the designated profile for interactive vide) ,
that means:
no B-Frames (too much latency in interactive video)
CBR (we want to get the max. quality making use of all the bitrate that is available)
We allow one exeption: configuring a bitrate of > 786 kbit/s

This plugin implements
- Single Time Aggregation Packets A
- Single NAL units
- Fragmentation Units
like described in RFC3984

It requires x264 and ffmpeg.


git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@17829 023b2edf-31b2-4de3-b41e-bca80c47788f