After the removal (due to multithreading issues) of the PCollection caching of integer based offsets (operator[]), many PList variables in the system have become hideously inefficient. Two things were done to fix this:
Convert any PList variable that could be to a PArray variable. This was quite a few as a lot do not do insertions or deletions from within the list, only append at the start.

Utilise new STL-like iterators added to the PList class.

Note the old operator[] is still available and will work as before, it is just not desirable to use it.