Fixed OpalCall party names being set to correct values, including URL like protocol prefix. They should then be usable for "last party redial" functions etc.
Added function to OpalConnection and derived classes to return indication of a new concept, a "network" connection. This is best described by example: sip, h323, iax and pstn are all "network" connections as they connect to something "remote". While pc, pots and ivr are not as the entity being connected to is intrinsically local.

This makes coding some aspects of softphone applications much simpler.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@20173 023b2edf-31b2-4de3-b41e-bca80c47788f