Major rewrite of fax.
Now utilises the latest SpanDSP, and since it's change to LGPL, it uses it as a DLL instead of having to talk to another process down a socket. All three fax modes are now in a standard codec plug in. When sending files a pseudo-format of "TIFF-File" is used and a codec from "TIFF-File" to/from "T.38" does the translation. Similarly there is "TIFF-File" to/from "PCM-16" (and thus on to G.711) and "PCM-16" to/from "T.38" for physical fax machines on a LID.

Note this is first cut and not every mode is tested yet.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@22200 023b2edf-31b2-4de3-b41e-bca80c47788f