Implemented "internal transfer" so swap the "back end" connection, for example SIP/H.323 call from PCSS to IVR and back again.
In the process changed MakeConnection() function return a PSafePtr to the connection made, which prevents some race conditions and extra logic to get that pointer after the call returns.

Similarly, an overloaded OpalManager::SetUpCall() was added to return a OpalCall PSafePtr.

Also, added ability to have ;params to a PCSS endpoint URI.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23008 023b2edf-31b2-4de3-b41e-bca80c47788f