Fixed losing the timestamp from silence insertion packets received from jitter buffer when packets are too late or absent.
Added "frame size" as well as the total data buffering size to the OpalMediaStream SetDataSize() function. Only really used for audio codecs and helping with the real time aspect of the audio sound device media stream. Essentially there needs to be a distinction between the sound card buffering (based on frame size) and the maximum packet size which is frameSize*maxRxFrames.


git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23036 023b2edf-31b2-4de3-b41e-bca80c47788f