Added ability for video decoder plug in to indicate that the output buffer needs to be resized. Useful for things like H.263 where it could change from QCIF to CIF on the fly and require a larger YUV frame.
Optimised the video output to re-use the previously allocated output frame. As video frames can be large, when the heap gets a bit fragmented it slows the system down substantially to search for a large enough free memory block, so as we don't have to make a new one every time, let's not.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@23253 023b2edf-31b2-4de3-b41e-bca80c47788f