Fixed issues when starting SIP calls "in hold" and an apparently simple problem exploded into a multitude of changes.
First the system is made much more consistent about the presence/absence of media stream associated with corresponding SDP m= lines. Now, if m= is present and port is non zero there will be media streams and media patch, but may be "paused" or "held" due the recvonly/sendonly/inactive state. Only if the m= is absent or it's port number is zero, or c= line is 0.0.0.0 will the media streams be closed and media patch shut down.

Second the Hold/Retrieve system to be changed so you can request a remote hold to be lifted. Previously only local holds (us putting them on hold) was controllable.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@24244 023b2edf-31b2-4de3-b41e-bca80c47788f