Possible fix for bug ID: 3495599 "Dead lock on DTMF while clearing call", reported by Valeriy V. Argunov.
Rearranged the OnReleased thread for cleaning up connections. Old system was prone to deadlocks depending on where the OpalConection::Release() was called. But we still don't want to spawn a thread for every OnReleased() function no matter how trivial. New system tries to have it both ways by making sure the last connection released in a call is in the same thread as the previous connects. As 99.99% of calls have exactly 2 connections, this does the trick. The remaining rare cases just create more threads and will still work.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/branches/v3_10@27078 023b2edf-31b2-4de3-b41e-bca80c47788f