Added full RFC5626 (NAT traversal) support, and partial RFC5923 (connection reuse) and better compliance to RFC5922 (domain certificates).
The primary requirement of the above is the ability for a transport connection (e.g. a single TLS link) to be used for multiple SIP entities, in particular a registration and subsequent calls using that registrar/proxy. It also applies to any SUBSCRIBE requests for presence etc. OPAL did not really support this and would create separate transport connections for each, with the one exception of and earlier hack that had the REGISTER and INVITE using the same local UDP socket.

Unfortunately, this seemingly simple thing has required the uprooting, rotating, folding, spindling and general mangling of everything the SIP code was doing with transports. Regressions for corner cases are likely.

git-svn-id: svn://svn.code.sf.net/p/opalvoip/code/opal/trunk@28394 023b2edf-31b2-4de3-b41e-bca80c47788f